refactor : LottoNumberList 클래스 삭제, 테스트 코드 수정
