+++ src/test/java/lotto/domain/LottoNumberListTest.java
@@ -10,10 +10,10 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 public class LottoNumberListTest {
 
   @Test
-  @DisplayName("지난주 로또에 이번 주 로또가 얼마나 포함되는지")
+  @DisplayName("지난주 로또에 이번 주 로또가 몇 개나 포함되는지")
   public void containsCount() {
-    LottoNumberList lottoNumberList = LottoNumberList.of(Arrays.asList(1, 2, 3, 4, 5, 6));
+    Lotto lotto = Lotto.of(Arrays.asList(1, 2, 3, 4, 5, 6));
     LastWinningLotto lastWeekLotto = LastWinningLotto.of(Arrays.asList(1, 2, 3, 4, 5, 6), 4);
-    assertEquals(6, lottoNumberList.containsCount(lastWeekLotto));
+    assertEquals(6, lotto.containsCount(lastWeekLotto));
   }
 }
